Add aggregation rules for propagator metric (#59)
Since this metric can have a very large number of time series (one for
every distributed policy), we want to provide a way to store some
aggregations of it in ACM observabillity. These rules make what we
expect to be common requests: total and noncompliant counts by both
cluster and policy.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/16198
 - https://github.com/open-cluster-management/multicluster-observability-operator/pull/789

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>